working on redux
