+++ src/store/DeviceReducer.js
@@ -1,0 +1,9 @@
+import {deviceList} from './MockDevices'
+
+export const deviceReducer = (state = deviceList, action) => {
+    switch (action.state) {
+        default:
+            return state;
+    }
+}
+
