+++ src/store/Reducers.js
@@ -1,0 +1,7 @@
+import {combineReducers} from "redux";
+
+import {deviceReducer} from "./DeviceReducer.js";
+
+export const rootReducer = combineReducers({
+    devices: deviceReducer,
+});
